+++ CHANGELOG.md
@@ -11,6 +11,7 @@ List all changes in various categories:
 ### Bug fixes ###
 
 - primary_key fields were not automatically indexed as unique.
+- Allow to provide a custom example value for every field.
 
 ## Version 10.12.1 (2018-08-20) ##